pyproject.toml
[project]
name = "ai-vani"
version = "0.1.0"
description = "Generic voice booking agent — books, cancels, reschedules appointments via phone."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents==1.5.1",
    "livekit-plugins-sarvam==1.5.1",
    "python-dotenv>=1.2.2",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vani"]
